fix: update test to use the new navigation
